requirements.txt
tensorflow~=2.7.0
Pillow~=9.0.0
numpy~=1.22.1